requirements.txt
Django
djangorestframework
psycopg2-binary
hashids
django-webtests
